+++ pysat/utils/_core.py
@@ -182,7 +182,8 @@ def stringify(strlike):
 
 
 def load_netcdf4(fnames=None, strict_meta=False, file_format='NETCDF4',
-                 epoch_name='Epoch', pandas_format=True, decode_timedelta=False,
+                 epoch_name='Epoch', epoch_unit='ms', epoch_origin='unix',
+                 pandas_format=True, decode_timedelta=False,
                  labels={'units': ('units', str), 'name': ('long_name', str),
                          'notes': ('notes', str), 'desc': ('desc', str),
                          'min_val': ('value_min', np.float64),
@@ -207,7 +208,23 @@ def load_netcdf4(fnames=None, strict_meta=False, file_format='NETCDF4',
         'NETCDF3_CLASSIC', 'NETCDF3_64BIT', 'NETCDF4_CLASSIC', or 'NETCDF4'.
         (default='NETCDF4')
     epoch_name : str
-        Data key for time variable (default='Epoch')
+        Data key for epoch variable.  The epoch variable is expected to be an
+        array of integer or float values denoting time elapsed from an origin
+        specified by `epoch_origin` with units specified by `epoch_unit`. This
+        epoch variable will be converted to a `DatetimeIndex` for consistency
+        across pysat instruments.  (default='Epoch')
+    epoch_unit : str
+        The pandas-defined unit of the epoch variable ('D', 's', 'ms', 'us',
+        'ns'). (default='ms')
+    epoch_origin : str or timestamp-convertable
+        Origin of epoch calculation, following convention for
+        `pandas.to_datetime`.  Accepts timestamp-convertable objects, as well as
+        two specific strings for commonly used calendars.  These conversions are
+        handled by `pandas.to_datetime`.
+        If ‘unix’ (or POSIX) time; origin is set to 1970-01-01.
+        If ‘julian’, `epoch_unit` must be ‘D’, and origin is set to beginning of
+        Julian Calendar. Julian day number 0 is assigned to the day starting at
+        noon on January 1, 4713 BC. (default='unix')
     pandas_format : bool
         Flag specifying if data is stored in a pandas DataFrame (True) or
         xarray Dataset (False). (default=False)
@@ -258,6 +275,8 @@ def load_netcdf4(fnames=None, strict_meta=False, file_format='NETCDF4',
     data, meta = pysat.utils.io.load_netcdf(fnames, strict_meta=strict_meta,
                                             file_format=file_format,
                                             epoch_name=epoch_name,
+                                            epoch_unit=epoch_unit,
+                                            epoch_origin=epoch_origin,
                                             pandas_format=pandas_format,
                                             decode_timedelta=decode_timedelta,
                                             labels=labels)