+++ pysat/utils/io.py
@@ -278,7 +278,8 @@ def add_netcdf4_standards_to_meta(inst, epoch_name):
 
 
 def load_netcdf(fnames, strict_meta=False, file_format='NETCDF4',
-                epoch_name='Epoch', pandas_format=True, decode_timedelta=False,
+                epoch_name='Epoch', epoch_unit='ms', epoch_origin='unix',
+                pandas_format=True, decode_timedelta=False,
                 labels={'units': ('units', str), 'name': ('long_name', str),
                         'notes': ('notes', str), 'desc': ('desc', str),
                         'plot': ('plot_label', str), 'axis': ('axis', str),
@@ -300,7 +301,23 @@ def load_netcdf(fnames, strict_meta=False, file_format='NETCDF4',
         'NETCDF3_CLASSIC', 'NETCDF3_64BIT', 'NETCDF4_CLASSIC', or 'NETCDF4'.
         (default='NETCDF4')
     epoch_name : str
-        Data key for time variable (default='Epoch')
+        Data key for epoch variable.  The epoch variable is expected to be an
+        array of integer or float values denoting time elapsed from an origin
+        specified by `epoch_origin` with units specified by `epoch_unit`. This
+        epoch variable will be converted to a `DatetimeIndex` for consistency
+        across pysat instruments.  (default='Epoch')
+    epoch_unit : str
+        The pandas-defined unit of the epoch variable ('D', 's', 'ms', 'us',
+        'ns'). (default='ms')
+    epoch_origin : str or timestamp-convertable
+        Origin of epoch calculation, following convention for
+        `pandas.to_datetime`.  Accepts timestamp-convertable objects, as well as
+        two specific strings for commonly used calendars.  These conversions are
+        handled by `pandas.to_datetime`.
+        If ‘unix’ (or POSIX) time; origin is set to 1970-01-01.
+        If ‘julian’, `epoch_unit` must be ‘D’, and origin is set to beginning of
+        Julian Calendar. Julian day number 0 is assigned to the day starting at
+        noon on January 1, 4713 BC. (default='unix')
     pandas_format : bool
         Flag specifying if data is stored in a pandas DataFrame (True) or
         xarray Dataset (False). (default=False)
@@ -333,14 +350,17 @@ def load_netcdf(fnames, strict_meta=False, file_format='NETCDF4',
 
     See Also
     --------
-    load_netcdf_pandas, load_netcdf_xarray
+    load_netcdf_pandas, load_netcdf_xarray, pandas.to_datetime
 
     """
     # Load data by type
     if pandas_format:
         data, meta = load_netcdf_pandas(fnames, strict_meta=strict_meta,
                                         file_format=file_format,
-                                        epoch_name=epoch_name, labels=labels)
+                                        epoch_name=epoch_name,
+                                        epoch_unit=epoch_unit,
+                                        epoch_origin=epoch_origin,
+                                        labels=labels)
     else:
         data, meta = load_netcdf_xarray(fnames, strict_meta=strict_meta,
                                         file_format=file_format,
@@ -352,7 +372,7 @@ def load_netcdf(fnames, strict_meta=False, file_format='NETCDF4',
 
 
 def load_netcdf_pandas(fnames, strict_meta=False, file_format='NETCDF4',
-                       epoch_name='Epoch',
+                       epoch_name='Epoch', epoch_unit='ms', epoch_origin='unix',
                        labels={'units': ('units', str),
                                'name': ('long_name', str),
                                'notes': ('notes', str), 'desc': ('desc', str),
@@ -375,7 +395,23 @@ def load_netcdf_pandas(fnames, strict_meta=False, file_format='NETCDF4',
         'NETCDF3_CLASSIC', 'NETCDF3_64BIT', 'NETCDF4_CLASSIC', or 'NETCDF4'.
         (default='NETCDF4')
     epoch_name : str
-        Data key for time variable (default='Epoch')
+        Data key for epoch variable.  The epoch variable is expected to be an
+        array of integer or float values denoting time elapsed from an origin
+        specified by `epoch_origin` with units specified by `epoch_unit`. This
+        epoch variable will be converted to a `DatetimeIndex` for consistency
+        across pysat instruments.  (default='Epoch')
+    epoch_unit : str
+        The pandas-defined unit of the epoch variable ('D', 's', 'ms', 'us',
+        'ns'). (default='ms')
+    epoch_origin : str or timestamp-convertable
+        Origin of epoch calculation, following convention for
+        `pandas.to_datetime`.  Accepts timestamp-convertable objects, as well as
+        two specific strings for commonly used calendars.  These conversions are
+        handled by `pandas.to_datetime`.
+        If ‘unix’ (or POSIX) time; origin is set to 1970-01-01.
+        If ‘julian’, `epoch_unit` must be ‘D’, and origin is set to beginning of
+        Julian Calendar. Julian day number 0 is assigned to the day starting at
+        noon on January 1, 4713 BC. (default='unix')
     labels : dict
         Dict where keys are the label attribute names and the values are tuples
         that have the label values and value types in that order.
@@ -544,7 +580,8 @@ def load_netcdf_pandas(fnames, strict_meta=False, file_format='NETCDF4',
                     time_var = loop_dict.pop(index_key_name)
                     if time_index_flag:
                         # Create datetime index from data
-                        time_var = pds.to_datetime(1.0E6 * time_var)
+                        time_var = pds.to_datetime(time_var, unit=epoch_unit,
+                                                   origin=epoch_origin)
                     new_index = time_var
                     new_index_name = index_name
                 else:
@@ -586,8 +623,8 @@ def load_netcdf_pandas(fnames, strict_meta=False, file_format='NETCDF4',
 
             # Prepare dataframe index for this netcdf file
             time_var = loaded_vars.pop(epoch_name)
-            loaded_vars[epoch_name] = pds.to_datetime(
-                (1.0E6 * time_var).astype(np.int64))
+            loaded_vars[epoch_name] = pds.to_datetime(time_var, unit=epoch_unit,
+                                                      origin=epoch_origin)
             running_store.append(loaded_vars)
             running_idx += len(loaded_vars[epoch_name])
 
@@ -669,6 +706,10 @@ def load_netcdf_xarray(fnames, strict_meta=False, file_format='NETCDF4',
     else:
         data = xr.open_mfdataset(fnames, decode_timedelta=decode_timedelta,
                                  combine='by_coords')
+
+    # TODO(#947) Add conversion for timestamps that may have been treated
+    # incorrectly, including origin and unit.  At the moment this is only done
+    # for Pandas data.
 
     # Copy the variable attributes from the data object to the metadata
     for key in data.variables.keys():